Move all API code to one file.
